UI folder create - scripts are moved
