--- SchoolRush/Assets/Scripts/SaveNickname.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c989bdae902745b2b39caa0107a0e90
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
