--- SchoolRush/Assets/Scripts/SceneManagerCS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd2b92666738648a394c7fa505fb23fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
