+++ SchoolRush/Assets/Scripts/UI/HelpManager.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class HelpManager : MonoBehaviour
+{
+    public GameObject HelpImage;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
