+++ SchoolRush/Assets/Scripts/UI/HelpManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2832d978e421f4cc3975a5b9c609df2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
