+++ SchoolRush/Assets/Scripts/UI/SceneManagerCS.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SceneManagerCS : MonoBehaviour {
+    public void loadScene(string sceneName) {
+        UnityEngine.SceneManagement.SceneManager.LoadScene(sceneName);
+    }
+}
