--- SchoolRush/Assets/Scripts/WebManagerCS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2936c14bd730e4e65b98d1ecfbb346fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
